fixed an initialization errror in slapt_is_required_by() that caused a segfault when removing with --no-deps.
if no package sources are defined with a priority, disable use of priorities when dealing with /patches.  This prevents the previous default slapt-get behavior from changing and causing confusion.  With priorities, /patches is always assigned +1 above the package sources priority.  So if no priorities were in use then anything found in any sources /patches always trumped anything else.  This is probably the most correct behavior but it is still unexpected for existing users.  To get the correct behavior, assign a priority to one or more sources.